changes for filter and sorting.

+++ src/app/card/card.component.ts
@@ -9,7 +9,9 @@ import { SnackbarService } from "ngx-snackbar";
 })
 export class CardComponent implements OnInit {
   @Input() users: any
+  filterData: any;
   tasks: any = [];
+  searchRecords = [];
   newTask: any;
   constructor(private taskSrc: TaskService, private snackbarService: SnackbarService) { }
 
@@ -28,15 +30,13 @@ export class CardComponent implements OnInit {
     this.taskSrc.getTaskList().subscribe(data => {
       this.tasks = data.tasks;
       this.tasks.map(x => x.due_date = moment(x.due_date).format('DD-MMM-YYYY'));
+      this.tasks = this.tasks.sort((d1, d2) => new Date(d1.due_date).getTime() - new Date(d2.due_date).getTime());
+      this.searchRecords = this.tasks;
     });
   }
 
   addCard() {
     this.tasks.push(this.newTask)
-  }
-
-  setTaskDate(event) {
-    console.log(event)
   }
 
   addNewTask(task) {
@@ -46,6 +46,7 @@ export class CardComponent implements OnInit {
     this.taskSrc.createTask(formData).subscribe(res => {
       if (res.status == "success" && res.taskid) {
         task.id = res.taskid
+        task.assigned_name = this.users.filter(x => x.id == task.assigned_name)[0];
         this.snackbarService.add({
           msg: 'Record Added Successfully.',
           timeout: 3000,
@@ -62,23 +63,25 @@ export class CardComponent implements OnInit {
 
   deleteTask(id) {
     if (id) {
-      let formData = new FormData();
-      formData.append("taskid", id);
-      this.taskSrc.deleteTask(formData).subscribe(res => {
-        if (res.status == "success") {
-          this.tasks = this.tasks.filter(x => x.id != id);
-          this.snackbarService.add({
-            msg: 'Record Deleted Successfully.',
-            timeout: 3000,
-            action: {
-              text: '',
-              onClick: (snack) => {
-                console.log('dismissed: ' + snack.id);
-              },
-            }
-          });
-        }
-      })
+      if (confirm("Are you sure, You wan't to delete this record?")) {
+        let formData = new FormData();
+        formData.append("taskid", id);
+        this.taskSrc.deleteTask(formData).subscribe(res => {
+          if (res.status == "success") {
+            this.tasks = this.tasks.filter(x => x.id != id);
+            this.snackbarService.add({
+              msg: 'Record Deleted Successfully.',
+              timeout: 3000,
+              action: {
+                text: '',
+                onClick: (snack) => {
+                  console.log('dismissed: ' + snack.id);
+                },
+              }
+            });
+          }
+        })
+      }
     }
   }
 
@@ -107,5 +110,10 @@ export class CardComponent implements OnInit {
         });
       }
     })
+  }
+
+  filterRecords() {
+    this.tasks = this.searchRecords;
+    this.tasks = this.tasks.filter(x => x.assigned_name.toLowerCase().indexOf(this.filterData.toLowerCase()) != -1)
   }
 }
